+++ .github/workflows/main.yml
@@ -19,6 +19,8 @@ jobs:
           toolchain: ${{ matrix.rust }}
           profile: minimal
           override: true
+      - name: Clippy
+        run: cargo clippy --all-targets --all-features -- -D warnings
       - name: Install 3proxy
         run: |
           cd $HOME
@@ -27,9 +29,13 @@ jobs:
           cd 3proxy-0.8.13 && ln -s Makefile.Linux Makefile && make -j$(nproc)
           sudo apt-get update
           sudo apt-get install socat -y
-      - name: Run tests
+      - name: Run tests for tokio
         run: |
           cargo build --examples
           cargo build --verbose --all
+          cargo test --lib --verbose
+          env PATH=$HOME/3proxy-0.8.13/src:$PATH tests/integration_tests.sh
+      - name: Run tests for futures
+        run: |
           cargo test --lib --verbose
           env PATH=$HOME/3proxy-0.8.13/src:$PATH tests/integration_tests.sh
