+++ Cargo.toml
@@ -38,7 +38,9 @@ futures_compat = { package = "futures", version = "0.3", default-features = fals
 ], optional = true }
 
 [dev-dependencies]
+test-utils = { path = "test-utils" }
 futures-executor = "0.3"
 tokio = { version = "1.0", features = ["io-util", "rt-multi-thread", "net"] }
+smol = "2.0"
 once_cell = "1.2.0"
 hyper = "0.14"
