+++ test-utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "test-utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+futures-executor = "0.3"
+tokio = { version = "1.0", features = ["io-util", "rt-multi-thread", "net"] }
+smol = "2.0"
+once_cell = "1.2.0"
+hyper = "0.14"
+tokio-socks = { path = ".." }
